fix:date index때문에 밀림현상
