+++ app/(home)/calendar/page.tsx
@@ -1,3 +1,4 @@
+import Calendar from '@/components/Calendar';
 import CalendarForm from '@/components/Calendar';
 import Memo from '@/components/Input/memo/Memo';
 import React from 'react';
@@ -5,7 +6,7 @@ import React from 'react';
 const CalendarPage = () => {
   return (
     <div>
-      <CalendarForm />
+      <Calendar />
     </div>
   );
 };
