+++ app/hooks/useCalendar.ts
@@ -5,23 +5,25 @@ import {
   startOfMonth,
   endOfMonth,
   eachDayOfInterval,
-  addMonths,
 } from 'date-fns';
 
 // useCalendar 훅의 반환 타입을 정의합니다.
 interface UseCalendarReturn {
   weeks: Date[][];
+  startDay: Date;
+  endDay: Date;
 }
 
 const useCalendar = (year: number, month: number): UseCalendarReturn => {
   const [weeks, setWeeks] = useState<Date[][]>([]);
-
+  const currentDate = new Date(year, month - 1, 1);
+  const startDay = startOfWeek(startOfMonth(currentDate));
+  const endDay = endOfWeek(endOfMonth(currentDate));
   useEffect(() => {
     // 현재 선택된 년과 월로 날짜 객체를 생성합니다.
-    // 월은 0부터 시작하므로, month - 1을 해줍니다.
-    const currentDate = new Date(year, month, 1);
-    const startDay = startOfWeek(startOfMonth(currentDate));
-    const endDay = endOfWeek(endOfMonth(currentDate));
+    // 월은 0부터 시작하므로, month에 -1을 해줍니다.
+    console.log(startDay, endDay);
+
     const days = eachDayOfInterval({ start: startDay, end: endDay });
 
     // 주별로 날짜를 분할합니다.
@@ -40,7 +42,7 @@ const useCalendar = (year: number, month: number): UseCalendarReturn => {
     setWeeks(newWeeks);
   }, [year, month]);
 
-  return { weeks };
+  return { weeks, startDay, endDay };
 };
 
 export default useCalendar;
