+++ app/recoil/calendar/atoms.ts
@@ -4,7 +4,7 @@ export const selectedDateState = atom({
   key: 'selectedDateState',
   default: {
     selectedYear: new Date().getFullYear(),
-    selectedMonth: new Date().getMonth(),
+    selectedMonth: new Date().getMonth() + 1,
     selectedDay: new Date().getDay(),
   },
 });
